net: keep track and avoid access of skb containing dma-buf pages.
Add checks through the network stack to avoid kmap()ing content
of the frags of devmem skbs.

Technically, whether or not data can be coalesced is a matter of
individual frag pages more than of the skb. An skb can combine
regular pages and dmabuf pages in principle, but in this current
implementation the skb->devmem flag either indicates that _all_ the
pages are devmem pages or none of them are. In the future skbs with
mixed page types can be added.

Change-Id: I9059a90935a19a41af3f9bd00a5ba0f19e9dc52d