[not-for-review] gve: unlink old napi when stopping a queue using queue API
When a queue is stopped using the ndo queue API, the napi id
associated with its page pool needs to be unlinked so that
it can be reused when the queue is restarted and a new page
pool is created for it.

Handle this by calling page_pool_disable_direct_recycling()
when stopping a queue.

Fixes: ebdfae0d377b ("gve: adopt page pool for DQ RDA mode")
Reported-by: Mina Almasry <almasrymina@google.com>
Change-Id: I2c36e3bac29379e0a2bce56ce84fdd38dbca9d8f
Signed-off-by: Harshitha Ramamurthy <hramamurthy@google.com>